tools/IoT-Inspector/iot_inspector.py
#!/usr/bin/env python3
import argparse
import serial
import scapy.all as scapy
import bluetooth
import os
import time
from datetime import datetime
import requests
import paho.mqtt.client as mqtt
from coapthon.client.helperclient import HelperClient
from art.estimators.classification import SklearnClassifier  # Placeholder for AI testing
import numpy as np

class IoTInspector:
    def __init__(self, report_file="iot_inspector_report.txt"):
        self.report_file = report_file
        self.report = []

    def log(self, message):
        timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
        entry = f"[{timestamp}] {message}"
        print(entry)
        self.report.append(entry)

    def extract_firmware(self, port, baudrate=115200):
        """Extract firmware via UART."""
        try:
            with serial.Serial(port, baudrate, timeout=1) as ser:
                self.log(f"Connecting to {port} at {baudrate} baud")
                ser.write(b"dump_firmware\n")
                firmware = ser.read(1024)
                with open("firmware.bin", "wb") as f:
                    f.write(firmware)
                self.log("Firmware extracted to firmware.bin")
        except Exception as e:
            self.log(f"Firmware extraction failed: {e}")

    def sniff_network(self, interface):
        """Sniff network traffic for IoT protocols."""
        self.log(f"Sniffing traffic on {interface}")
        packets = scapy.sniff(iface=interface, count=10)
        for pkt in packets:
            if scapy.Raw in pkt:
                self.log(f"Packet captured: {pkt.summary()}")
        scapy.wrpcap("iot_traffic.pcap", packets)
        self.log("Traffic saved to iot_traffic.pcap")

    def scan_bluetooth(self):
        """Discover Bluetooth devices."""
        self.log("Scanning for Bluetooth devices")
        devices = bluetooth.discover_devices(lookup_names=True)
        for addr, name in devices:
            self.log(f"Found Bluetooth device: {name} ({addr})")

    def test_matter_protocol(self, interface):
        """Sniff and analyze Matter protocol traffic."""
        self.log(f"Sniffing Matter traffic on {interface}")
        # Placeholder: Extend Scapy for Matter packet dissection
        packets = scapy.sniff(iface=interface, count=5)
        self.log("Matter protocol analysis complete")

    def test_lorawan(self, interface):
        """Test LoRaWAN vulnerabilities."""
        self.log(f"Testing LoRaWAN on {interface}")
        # Placeholder: Simulate join procedure and replay attacks
        self.log("LoRaWAN analysis complete")

    def test_5g_sidelink(self, interface):
        """Test 5G NR sidelink vulnerabilities."""
        self.log(f"Testing 5G NR sidelink on {interface}")
        # Placeholder: Use OpenAirInterface or similar for simulation
        self.log("5G NR sidelink analysis complete")

    def test_ocf(self, host, port=5683):
        """Test OCF protocol vulnerabilities."""
        self.log(f"Testing OCF on {host}:{port}")
        client = HelperClient(server=(host, port))
        response = client.get("/oic/res")
        self.log(f"OCF response: {response.pretty_print()}")
        client.stop()

    def test_ai_vulnerabilities(self):
        """Test for AI model vulnerabilities."""
        self.log("Scanning for AI model vulnerabilities")
        # Placeholder: Simulate adversarial attack with ART
        self.log("AI testing complete; check report for details")

    def test_zero_trust(self, host, port):
        """Test Zero Trust architecture."""
        self.log(f"Testing Zero Trust on {host}:{port}")
        try:
            response = requests.get(f"http://{host}:{port}", verify=False, timeout=5)
            self.log(f"Zero Trust response: {response.status_code}")
        except Exception as e:
            self.log(f"Zero Trust test failed: {e}")

    def test_satellite_connectivity(self, interface):
        """Test satellite IoT protocol resilience."""
        self.log(f"Simulating satellite connectivity on {interface}")
        # Placeholder: Emulate high-latency with NetEm or similar
        self.log("Satellite testing complete")

    def test_digital_twin(self, api_url):
        """Test digital twin security."""
        self.log(f"Testing digital twin at {api_url}")
        try:
            response = requests.get(api_url, timeout=5)
            self.log(f"Digital twin response: {response.status_code}")
        except Exception as e:
            self.log(f"Digital twin test failed: {e}")

    def test_quantum_cryptography(self):
        """Test for quantum-resistant cryptography."""
        self.log("Analyzing cryptography for quantum resistance")
        if os.path.exists("firmware.bin"):
            self.log("Checking firmware for weak algorithms (RSA, ECC)")
        if os.path.exists("iot_traffic.pcap"):
            self.log("Checking traffic for quantum-vulnerable encryption")
        self.log("Quantum cryptography analysis complete")

    def analyze_vulnerabilities(self):
        """Analyze firmware and traffic for vulnerabilities."""
        self.log("Analyzing firmware and traffic for vulnerabilities")
        if os.path.exists("firmware.bin"):
            self.log("Firmware file detected; checking for weak encryption")
        if os.path.exists("iot_traffic.pcap"):
            self.log("Traffic file detected; checking for unencrypted data")

    def generate_report(selfaporte):
        """Write report to file."""
        with open(self.report_file, "w") as f:
            f.write("\n".join(self.report))
        self.log(f"Report generated: {self.report_file}")

def main():
    parser = argparse.ArgumentParser(description="IoT-Inspector: IoT Penetration Testing Tool")
    parser.add_argument("--port", help="Serial port for firmware extraction (e.g., /dev/ttyUSB0)")
    parser.add_argument("--baudrate", type=int, default=115200, help="Baudrate for serial connection")
    parser.add_argument("--interface", help="Network interface for sniffing (e.g., wlan0)")
    parser.add_argument("--host", help="Host for protocol testing (e.g., 192.168.1.100)")
    parser.add_argument("--port-num", type=int, default=5683, help="Port for protocol testing")
    parser.add_argument("--api-url", help="API URL for digital twin testing")
    parser.add_argument("--test-ai", action="store_true", help="Test AI vulnerabilities")
    parser.add_argument("--test-zero-trust", action="store_true", help="Test Zero Trust architecture")
    parser.add_argument("--test-satellite", action="store_true", help="Test satellite connectivity")
    parser.add_argument("--test-digital-twin", action="store_true", help="Test digital twin security")
    parser.add_argument("--test-quantum", action="store_true", help="Test quantum cryptography")
    args = parser.parse_args()

    inspector = IoTInspector()

    if args.port:
        inspector.extract_firmware(args.port, args.baudrate)
    if args.interface:
        inspector.sniff_network(args.interface)
        inspector.test_matter_protocol(args.interface)
        inspector.test_lorawan(args.interface)
        inspector.test_5g_sidelink(args.interface)
        if args.test_satellite:
            inspector.test_satellite_connectivity(args.interface)
    inspector.scan_bluetooth()
    if args.host and args.port_num:
        inspector.test_ocf(args.host, args.port_num)
        if args.test_zero_trust:
            inspector.test_zero_trust(args.host, args.port_num)
    if args.test_ai:
        inspector.test_ai_vulnerabilities()
    if args.api_url and args.test_digital_twin:
        inspector.test_digital_twin(args.api_url)
    if args.test_quantum:
        inspector.test_quantum_cryptography()
    inspector.analyze_vulnerabilities()
    inspector.generate_report()

if __name__ == "__main__":
    main()
